requirements-test.txt
coverage
coveralls
mock>=1.0.1
pytest-django==2.8.0
flake8>=2.1.0
tox>=1.7.0
